Store domains as 32 bytes and overhaul receive native

+++ contracts/libraries/constraints/FlexChronologyConstraint.sol
@@ -7,7 +7,7 @@ import {FlexChronologyError} from "../../interfaces/errors/FlexChronologyError.s
 import {FlexSendStateAccess} from "../accesses/FlexSendStateAccess.sol";
 
 library FlexChronologyConstraint {
-    function validate(bytes32 bucketState_, uint48 start_) internal pure {
+    function validate(bytes32 bucketState_, uint256 start_) internal pure {
         require(FlexSendStateAccess.readTime(bucketState_) <= start_, FlexChronologyError());
     }
 }
