+++ contracts/libraries/constraints/FlexDeadlineConstraint.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.28;
 import {FlexDeadlineError} from "../../interfaces/errors/FlexDeadlineError.sol";
 
 library FlexDeadlineConstraint {
-    function validate(uint48 deadline_) internal view {
+    function validate(uint256 deadline_) internal view {
         require(block.timestamp <= deadline_, FlexDeadlineError());
     }
 }
