+++ contracts/libraries/constraints/FlexEarlinessConstraint.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.28;
 import {FlexEarlinessError} from "../../interfaces/errors/FlexEarlinessError.sol";
 
 library FlexEarlinessConstraint {
-    function validate(uint48 start_) internal view {
+    function validate(uint256 start_) internal view {
         require(block.timestamp >= start_, FlexEarlinessError());
     }
 }
